logo implemented

+++ Project/finance/lib/Providers/AnalysisManagement.dart
@@ -4,7 +4,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 import '../Entities/users.dart';
 
 class AnalysisManagement with ChangeNotifier, DiagnosticableTreeMixin {
-  Map<String, int> actualData15 = {};
-  Map<String, int> predictedData15 = {};
-  Map<String, int> lastMonthData30 = {};
+  Map<String, double> actualData15 = {};
+  Map<String, double> predictedData15 = {};
+  Map<String, double> lastMonthData30 = {};
 }
